refactor(sb): select csf3
